+++ src/Links/Tests/Dummies/TestLinkableMorphedProduct.php
@@ -1,0 +1,19 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Contains the TestLinkableMorphedProduct class.
+ *
+ * @copyright   Copyright (c) 2022 Attila Fulop
+ * @author      Attila Fulop
+ * @license     MIT
+ * @since       2022-12-01
+ *
+ */
+
+namespace Vanilo\Links\Tests\Dummies;
+
+class TestLinkableMorphedProduct extends TestLinkableProduct
+{
+}